TST: adds `pandas.util.testing.patch`
This function temporarily replaces a an attribute on an object for the
duration of some context manager. This is useful for patching methods to
inject assertions or mock out expensive operations.
